refactor: Delete unused viewport context
