+++ packages/insight-viewer/src/components/ViewerWrapper/index.tsx
@@ -1,9 +1,8 @@
 import React, { forwardRef } from 'react'
-import { WithChildren, Element, ProgressComponent } from '../../types'
+import { WithChildren, ProgressComponent } from '../../types'
 import LoadingProgress from '../LoadingProgress'
 import Wrapper from './Wrapper'
 import useResize from './useResize'
-import { ViewportContextProvider } from '../../Context/Viewport'
 
 const Forwarded = forwardRef<
   HTMLDivElement,
@@ -15,11 +14,7 @@ const Forwarded = forwardRef<
     <Wrapper ref={resizeRef}>
       {Progress && <LoadingProgress Progress={Progress} />}
       <canvas className="cornerstone-canvas" />
-      <ViewportContextProvider
-        element={(ref as React.MutableRefObject<Element>).current}
-      >
-        {children}
-      </ViewportContextProvider>
+      {children}
     </Wrapper>
   )
 })
