+++ packages/insight-viewer/src/index.ts
@@ -2,11 +2,10 @@ export { default } from './Viewer'
 export { default as useMultiframe } from './hooks/useMultiframe'
 export { default as useViewport } from './hooks/useViewport'
 export { default as useInteraction } from './hooks/useInteraction'
-export type { Viewport } from './Context/Viewport/types'
 export type {
   Interaction,
   SetInteraction,
   WheelEvent,
 } from './hooks/useInteraction/types'
-export type { ViewerError } from './types'
+export type { Viewport, ViewerError } from './types'
 export type { DragEvent, Drag } from './hooks/useInteraction/types'
